requirements.txt
django
gunicorn
psycopg2-binary==2.8.5
djangorestframework==3.11.1
python-decouple==3.3
docker-compose==1.28.2
drf-yasg==1.20.0
django-extensions==3.1.0
whitenoise==5.1.0
Pillow
django-bootstrap5
